+++ CHANGELOG.md
@@ -1,1 +1,5 @@
 ## 0.0.1 (SNAPSHOT)
+
+New:
+    - WLED API Controller.
+    - Toggle method.